not really anything

+++ utils/datamaker.py
@@ -4,6 +4,7 @@ import os
 import numpy as np
 from utils.Bbox import Bbox
 import cv2
+from utils import util
 
 
 
@@ -11,7 +12,7 @@ def get_data(config, root_dir="/dataset/"):
 
 	instance_count = 0
 	x_batch = np.zeros([config["BATCH_SIZE"], config["IMAGE_W"], config["IMAGE_H"], 3], np.float32)
-	anchors = [Bbox(0, 0, config["ANCHORS"][2*i], config["ANCHORS"][2*i + 1]) for i in range(len(config["ANCHORS"])/2)]
+	anchors = [Bbox(0, 0, config["ANCHORS"][2*i], config["ANCHORS"][2*i + 1]) for i in range(int(len(config["ANCHORS"])/2))]
 	y_batch = np.zeros([config["BATCH_SIZE"], config["GRID_W"], config["GRID_H"], config["BOX"], 4+1+config["CLASS"]], np.float32)
 	max_iou = -1
 	best_prior = -1
